Weird things happen using node_free in list_free (node_free gets called with the same value passe eventhough the value changes in the loop)

+++ src/monitor.h
@@ -18,7 +18,7 @@
  * 
  */
 typedef struct monitor {
-    short int id;
+    int id;
     int temp;
     char *read_path;
 } t_monitor;
